Add configurable playlist limits

+++ mpdfavd/playlistgenerator.go
@@ -15,7 +15,7 @@ func ListenSongStickerChange(songStickerChange chan SongSticker, handler songSti
 	}
 }
 
-func generatePlaylist(mpdc *MPDClient, stickerName string, playlistName string, max int, descending bool) {
+func generatePlaylist(mpdc *MPDClient, stickerName string, playlistName string, max uint, descending bool) {
 	log.Printf("playlist generator: regenerating %s\n", playlistName)
 	err := mpdc.PlaylistClear(playlistName)
 	if err != nil {
@@ -32,20 +32,20 @@ func generatePlaylist(mpdc *MPDClient, stickerName string, playlistName string, 
 			continue
 		}
 		mpdc.PlaylistAdd(playlistName, songSticker.Uri)
-		if i >= max {
+		if uint(i) >= max {
 			break
 		}
 	}
 }
 
-func generateBestRatedSongs(mpdc *MPDClient, playlistName string, max int) songStickerChangeHandler {
+func generateBestRatedSongs(mpdc *MPDClient, playlistName string, max uint) songStickerChangeHandler {
 	f := func(songSticker SongSticker) {
 		generatePlaylist(mpdc, songSticker.Name, playlistName, max, true)
 	}
 	return songStickerChangeHandler(f)
 }
 
-func generateMostPlayedSongs(mpdc *MPDClient, playlistName string, max int) songStickerChangeHandler {
+func generateMostPlayedSongs(mpdc *MPDClient, playlistName string, max uint) songStickerChangeHandler {
 	f := func(songSticker SongSticker) {
 		generatePlaylist(mpdc, songSticker.Name, playlistName, max, true)
 	}
